Add user logic

+++ src/user/user.controller.ts
@@ -14,7 +14,6 @@ import {
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { UserEntity } from './user.entity';
 import { UserService } from './user.service';
 import { StatusCodes } from 'http-status-codes';
 import { AppService } from 'src/app.service';
@@ -27,16 +26,16 @@ export class UserController {
   ) {}
 
   @Get()
-  findAllUsers(): UserEntity[] {
-    return this.userService.findAllUsers();
+  findAllUsers() {
+    return this.userService.getUsers();
   }
 
   @Get(':id')
-  findUserById(@Param('id') id: string): UserEntity {
+  async findUserById(@Param('id') id: string) {
     if (!this.appService.isValidUuid(id)) {
       throw new HttpException('Invalid id', StatusCodes.BAD_REQUEST);
     }
-    const user = this.userService.findUserById(id);
+    const user = await this.userService.getUserById(id);
     if (!user) {
       throw new NotFoundException('User not found');
     }
@@ -44,19 +43,19 @@ export class UserController {
   }
 
   @Post()
-  createUser(@Body() createUserDto: CreateUserDto): Partial<UserEntity> {
+  createUser(@Body() createUserDto: CreateUserDto) {
     return this.userService.createUser(createUserDto);
   }
 
   @Put(':id')
-  updateUser(
+  async updateUser(
     @Param('id') id: string,
     @Body() updateUserDto: UpdateUserDto,
-  ): Partial<UserEntity> {
+  ) {
     if (!this.appService.isValidUuid(id)) {
       throw new HttpException('Invalid user ID', HttpStatus.BAD_REQUEST);
     }
-    const user = this.findUserById(id);
+    const user = await this.userService.getUserById(id);
     if (!user) {
       throw new NotFoundException('User not found');
     }
@@ -70,10 +69,14 @@ export class UserController {
 
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  deleteUser(@Param('id') id: string): void {
+  async deleteUser(@Param('id') id: string): Promise<void> {
     if (!this.appService.isValidUuid(id)) {
       throw new HttpException('Invalid id', StatusCodes.BAD_REQUEST);
     }
-    return this.userService.deleteUser(id);
+    const user = await this.userService.getUserById(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    this.userService.deleteUser(id);
   }
 }
